While profiling an application using `roslibrust` that relies on this crate for serialization logic. I discovered that the application was taking a substantial amount of time to parse an image message using serde_rosmsg.

Diving in I discovered that even with an optimized release build, the code was ending up walking every byte in the image resulting in ~2.3 million function calls to parse a 1080p image.

After a decent amount of fiddling and trying different solutions, I discovered that https://docs.rs/serde_bytes/latest/serde_bytes/ could be used to cue serde into calling deserialize_byte_buf for the case of Deserializing into a Vec<u8> and that by specializing that  case to bulk deserialize the entire array, we could turn an iteration over every byte in an image to effectively a bulk memcpy.

Comparing the code that hits the optimization to the code that doesn't:
```
parse_vec_image         time:   [22.450 ms 22.543 ms 22.643 ms]
                        change: [-2.4723% -2.0143% -1.5825%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high severe

parse_vec_bytes_image   time:   [527.16 µs 528.77 µs 530.31 µs]
                        change: [-4.6984% -3.1317% -1.7029%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe
  ```
  
We go from 22.5ms to deserialize a 1080p image to 530us :partying_face: :tada:.
  
This is a relatively niche optimization as it only applies to uint8[] to Vec<u8> conversions, but as this is the underlying buffer type or sensor_msgs/Image and sensor_msgs/PointCloud2, this will mean a massive speed up for some common ROS use cases.

I've included the benchmarking code in this MR in case that is useful, but would be happy to prepare a version with only the minimal set of changes.

 For reference here is the flamegraph from the version that misses this optimization: 
![flamegraph](https://github.com/user-attachments/assets/041765c3-af39-4a04-b394-b5de55dafcc8)
You can see bulk of time being spent in VecVisitor<T> individually accessing each element.

And here is with the optimization:
![flamegraph](https://github.com/user-attachments/assets/de4746d6-d162-4576-8e46-201524bedf1b)
You can see that bulk of time is spent in allocation, and AVX copies. In this version ~50% of time is being spent in allocation, so the only remaining optimization is to get deserialization to `&'de [u8]` working.